requirements.txt
Flask>=2.0
gunicorn
psycopg2-binary
python-dotenv
razorpay